+++ rq/job.py
@@ -810,7 +810,7 @@ class Job:
         return self._exc_info
 
     def return_value(self, refresh: bool = False) -> Optional[Any]:
-        """Returns the return value of the latest execution, if it was successful
+        """Returns the return value of the latest execution, if it was successful.
 
         Args:
             refresh (bool, optional): Whether to refresh the current status. Defaults to False.
@@ -889,16 +889,18 @@ class Job:
 
         return Result.all(self, serializer=self.serializer)
 
-    def latest_result(self) -> Optional['Result']:
+    def latest_result(self, timeout: int = 0) -> Optional['Result']:
         """Get the latest job result.
+
+        Args:
+            timeout (int, optional): Number of seconds to block waiting for a result. Defaults to 0 (no blocking).
 
         Returns:
             result (Result): The Result object
         """
-        """Returns the latest Result object"""
         from .results import Result
 
-        return Result.fetch_latest(self, serializer=self.serializer)
+        return Result.fetch_latest(self, serializer=self.serializer, timeout=timeout)
 
     def restore(self, raw_data) -> Any:
         """Overwrite properties with the provided values stored in Redis.
@@ -1607,7 +1609,7 @@ class Job:
             # If parent job is not finished, we should only continue
             # if this job allows parent job to fail
             dependencies_ids.discard(parent_job.id)
-            if parent_job._status == JobStatus.CANCELED:
+            if parent_job.get_status() == JobStatus.CANCELED:
                 return False
             elif parent_job._status == JobStatus.FAILED and not self.allow_dependency_failures:
                 return False